+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/operations/upload_attachment.py
@@ -3,17 +3,9 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import attachment as shared_attachment
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import attachment as shared_attachment
+from ...models.shared import attachmentupload as shared_attachmentupload
 from typing import Optional
-
-
-@dataclasses.dataclass
-class UploadAttachmentRequestBody:
-    content: bytes = dataclasses.field(metadata={'multipart_form': { 'content': True }})
-    request_body: str = dataclasses.field(metadata={'multipart_form': { 'field_name': 'requestBody' }})
-    
-
 
 
 @dataclasses.dataclass
@@ -24,7 +16,7 @@ class UploadAttachmentRequest:
     r"""Unique identifier for a sync."""
     transaction_id: str = dataclasses.field(metadata={'path_param': { 'field_name': 'transactionId', 'style': 'simple', 'explode': False }})
     r"""The unique identifier for your SMB's transaction."""
-    request_body: Optional[UploadAttachmentRequestBody] = dataclasses.field(default=None, metadata={'multipart_form': { 'file': True }, 'request': { 'media_type': 'multipart/form-data' }})
+    attachment_upload: Optional[shared_attachmentupload.AttachmentUpload] = dataclasses.field(default=None, metadata={'request': { 'media_type': 'multipart/form-data' }})
     
 
 
@@ -33,13 +25,11 @@ class UploadAttachmentRequest:
 class UploadAttachmentResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     attachment: Optional[shared_attachment.Attachment] = dataclasses.field(default=None)
     r"""OK"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""The request made is not valid."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""