+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/shared/__init__.py
@@ -2,7 +2,9 @@
 
 from .accountmappinginfo import *
 from .attachment import *
+from .attachmentupload import *
 from .bankaccount import *
+from .codatfile import *
 from .companies import *
 from .company import *
 from .companyconfiguration import *
@@ -16,7 +18,6 @@ from .createexpenseresponse import *
 from .customer import *
 from .dataconnectionerror import *
 from .dataconnectionstatus import *
-from .errormessage import *
 from .expensetransaction import *
 from .expensetransactionline import *
 from .halref import *
@@ -31,7 +32,6 @@ from .synccompletewebhook import *
 from .synccompletewebhookdata import *
 from .syncfailedwebhook import *
 from .syncfailedwebhookdata import *
-from .syncinitiated import *
 from .syncstartedwebhook import *
 from .taxratemappinginfo import *
 from .trackingcategorymappinginfo import *
@@ -40,4 +40,4 @@ from .transactionmetadatalist import *
 from .transactionstatus import *
 from .updateexpenserequest import *
 
-__all__ = ["AccountMappingInfo","AccountMappingInfoAccountType","AccountMappingInfoValidTransactionTypes","Attachment","BankAccount","Companies","Company","CompanyConfiguration","CompanyRequestBody","CompanySyncStatus","Connection","ConnectionSourceType","Connections","ContactRef","ContactRefContactType","CreateExpenseRequest","CreateExpenseResponse","Customer","DataConnectionError","DataConnectionStatus","ErrorMessage","ExpenseTransaction","ExpenseTransactionLine","ExpenseTransactionType","HalRef","IntegrationType","Links","MappingOptions","PostSync","RecordRef","Security","Supplier","SyncCompleteWebhook","SyncCompleteWebhookData","SyncFailedWebhook","SyncFailedWebhookData","SyncInitiated","SyncStartedWebhook","SyncStartedWebhookData","TaxRateMappingInfo","TaxRateMappingInfoValidTransactionTypes","TrackingCategoryMappingInfo","TransactionMetadata","TransactionMetadataList","TransactionStatus","UpdateExpenseRequest"]
+__all__ = ["AccountMappingInfo","AccountType","Attachment","AttachmentUpload","BankAccount","CodatFile","Companies","Company","CompanyConfiguration","CompanyRequestBody","CompanySyncStatus","Connection","Connections","ContactRef","ContactType","CreateExpenseRequest","CreateExpenseResponse","Customer","Data","DataConnectionError","DataConnectionStatus","ExpenseTransaction","ExpenseTransactionLine","HalRef","IntegrationType","Links","MappingOptions","PostSync","RecordRef","Security","SourceType","Supplier","SyncCompleteWebhook","SyncCompleteWebhookData","SyncFailedWebhook","SyncFailedWebhookData","SyncStartedWebhook","TaxRateMappingInfo","TaxRateMappingInfoValidTransactionTypes","TrackingCategoryMappingInfo","TransactionMetadata","TransactionMetadataList","TransactionStatus","Type","UpdateExpenseRequest","ValidTransactionTypes"]