+++ previous-versions/sync-for-expenses-version-1/src/codatsyncexpenses/models/shared/contactref.py
@@ -7,7 +7,7 @@ from dataclasses_json import Undefined, dataclass_json
 from enum import Enum
 from typing import Optional
 
-class ContactRefContactType(str, Enum):
+class ContactType(str, Enum):
     r"""The type of contact."""
     SUPPLIER = 'Supplier'
 
@@ -15,7 +15,7 @@ class ContactRefContactType(str, Enum):
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class ContactRef:
-    contact_type: Optional[ContactRefContactType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('contactType'), 'exclude': lambda f: f is None }})
+    contact_type: Optional[ContactType] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('contactType'), 'exclude': lambda f: f is None }})
     r"""The type of contact."""
     id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('id'), 'exclude': lambda f: f is None }})
     r"""Identifier of supplier or customer."""